requirements.txt
# dependencies
colorama
python-dotenv~=1.0.1
uvicorn~=0.34.0
pydantic~=2.10.4
fastapi~=0.115.6
aiokafka~=0.12.0
SQLAlchemy~=2.0.36
PyJWT
psycopg2~=2.9.10
redis
structlog
pydantic_settings
langchain_scrapegraph
python-dotenv
uvicorn
langchain~=0.3.14
langchain-openai
langsmith
motor~=3.6.0
pandas~=2.2.3
python-jobspy
starlette~=0.41.3
httpx~=0.28.1
jobspy~=0.29.0
asyncpg
langchain_openai
PyPDF2
langchain-community
pydantic[email]
python-multipart
passlib[bcrypt]
langgraph~=0.2.62
pypdf
langchain_anthropic
opencv-python~=4.11.0.86
PyMuPDF
pillow~=11.1.0
numpy~=1.26.3
typing_extensions~=4.12.2
langchain_google_vertexai
langgraph-cli[inmem]